% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate two data sets with some highly correlated variables}
\usage{
generate_data(number_of_ksi = 1, number_of_patients = 50,
  number_of_Xs_associated_with_ksis = c(5),
  number_of_not_associated_Xs = 250,
  mean_of_the_regression_weights_of_the_associated_Xs = c(0.7),
  sd_of_the_regression_weights_of_the_associated_Xs = c(0.05),
  Xnoise_min = -0.3, Xnoise_max = 0.3,
  number_of_Ys_associated_with_ksis = c(5),
  number_of_not_associated_Ys = 350,
  mean_of_the_regression_weights_of_the_associated_Ys = c(0.7),
  sd_of_the_regression_weights_of_the_associated_Ys = c(0.05),
  Ynoise_min = -0.3, Ynoise_max = 0.3)
}
\arguments{
\item{number_of_ksi}{The number of latent variables of the predictive data
set.}

\item{number_of_patients}{The number of observations (rows) in the data sets.}

\item{number_of_Xs_associated_with_ksis}{Number of variables of the
predictive data set that is associated with the predicted data set.}

\item{number_of_not_associated_Xs}{Number of variables of the predictive data
set that is not associated with the predicted data set.}

\item{mean_of_the_regression_weights_of_the_associated_Xs}{Mean of the
regression weights of the predictive varaibles that are associated with the
predicted variables.}

\item{sd_of_the_regression_weights_of_the_associated_Xs}{Standard deviation
of the regression weights of the predictive varaibles that are associated
with the predicted variables.}

\item{Xnoise_min}{The lower bound of the unifrom distribution that is used to
sample the values for the regression weights of the predictive varaibles that
are not associated with the predicted variables.}

\item{Xnoise_max}{The upper bound of the unifrom distribution that is used to
sample the values for the regression weights of the predictive varaibles that
are not associated with the predicted variables.}
}
\description{
Generate two data sets with some highly correlated variables.
}
\examples{
dataXY <- generate_data(number_of_ksi = 1,
                           number_of_patients = 500,
                           number_of_Xs_associated_with_ksis = c(5),
                           number_of_not_associated_Xs = 250,
                           mean_of_the_regression_weights_of_the_associated_Xs =
                             c(0.9),
                           sd_of_the_regression_weights_of_the_associated_Xs =
                             c(0.05),
                           Xnoise_min = -0.3, Xnoise_max = 0.3,
                           
                           number_of_Ys_associated_with_ksis = c(5),
                           number_of_not_associated_Ys = 350,
                           mean_of_the_regression_weights_of_the_associated_Ys =
                             c(0.9),
                           sd_of_the_regression_weights_of_the_associated_Ys =
                             c(0.05),
                           Ynoise_min = -0.3, Ynoise_max = 0.3)

# seperate predictor and predicted sets
X <- dataXY$X
Y <- dataXY$Y

dim(X);dim(Y)

}
\keyword{data}
\keyword{generate}
