% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dummies.r
\name{get_dummy}
\alias{get_dummy}
\title{Non-Metric Dummy}
\usage{
get_dummy(x)
}
\arguments{
\item{x}{a vector whose elements are non-negative integers from 0 to p}
}
\value{
the dummy matrix
}
\description{
Internal function. \code{get_dummy} is called by \code{get_dummies}.
Transforms a vector of natural numbers from 1 to p into the
corresponding p x p dummy matrix
}
\keyword{internal}
