% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enet2.R
\name{get_enet2}
\alias{get_enet2}
\title{Calculate elastic net}
\usage{
get_enet2(X.m, z, lambda, nonzero, alpha = 0.5)
}
\description{
Internal function not to be called by the user
}
\details{
Calculate elastic net with glmnet package
}
\section{Warning}{

  Do NOT use this function unless you are ME,
  a package developer, or a jedi user who really
  knows what is doing (seriously!)
}

\keyword{internal}
