% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plspm-package.r
\docType{package}
\name{plspm}
\alias{plspm}
\alias{plspm-package}
\title{\pkg{plspm} Tools for Partial Least Squares Path Modeling (PLS-PM)}
\description{
Estimate path models with latent variables by partial least squares 
approach (for both metric and non-metric data)
}
\details{
\itemize{
  \item \code{\link{plspm}}: Partial Least Squares Path Modeling
  \item \code{\link{plspm.fit}}: Simple version for PLS-PM
  \item \code{\link{plspm.groups}}: Two Groups Comparison in PLS-PM
  \item \code{\link{rebus.pls}}: Response Based Unit Segmentation (REBUS)
}
}
