% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/russett-data.r
\docType{data}
\name{russa}
\alias{russa}
\title{Russett A}
\format{A data frame with 47 rows and 9 columns}
\usage{
russa
}
\description{
Russett dataset with variable \code{demo} as numeric variable
}
\keyword{datasets}
