% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/russett-data.r
\docType{data}
\name{russb}
\alias{russb}
\title{Russett B}
\format{A data frame with 47 rows and 9 columns}
\usage{
russb
}
\description{
Russett dataset with variable \code{demo} as factor
}
\keyword{datasets}
